refactoring sqltomvt
